requirements.txt
numpy
requests
beautifulsoup4
mysql-connector-python
tqdm
pandas
logging  # nativa do Python, não precisa instalar
zipfile36  # só necessário em versões muito antigas do Python (<3.6)



DB_CONFIG = {
    "host": "localhost",
    "user": "root",
    "password": "33548084",
    "database": "cnpj_brasil"
}